Move the Param abstract class into the DockerClient file.
